Attempt to fix a potential race in e2e tests
In github e2e tests, it looks like it sometimes happens that the
headscale node list doesn't list the tsnsrv host, which then caused us
to take the 0th index of an empty list.

So to make future diagnosis easier, put the code into functions (which
should appear on the backtrace), and wait until we have exactly one
matching element.