Make nixos tests fail faster if tsnsrv unit fails
If the unit failed to start and never registered with headscale, we'd
wait until the test timeout to fail - but it would never be able to
succeed. Instead, let's use a polling condition that ensures the
tsnsrv unit is active & have the harness abort the test when that
exits, ever.