Allow specifying OAuth client secrets and tags on the CLI
When using tailscale OAuth2 clients, we must mint an auth key from
that client secret first. That requires passing tags to the
program (which must match what the client was minted with).

This allows granting long-lived (until the client is revoked) access
to tsnsrv, for authentication purposes (but it itself mints very
short-lived auth keys).